+++ pom.xml
@@ -4,11 +4,11 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.shapesecurity</groupId>
-    <artifactId>shift-bundler</artifactId>
-    <version>1.0.1</version>
+    <artifactId>bandolier</artifactId>
+    <version>1.0.0</version>
     <packaging>jar</packaging>
 
-    <name>Shift Bundler</name>
+    <name>Bandolier</name>
     <description>Bundles es2015 modules</description>
 
     <dependencies>
