+++ src/main/java/com/shapesecurity/bandolier/ModuleDependencyCollector.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler;
+package com.shapesecurity.bandolier;
 
 import com.shapesecurity.functional.data.ImmutableList;
 import com.shapesecurity.functional.data.Maybe;
