+++ src/main/java/com/shapesecurity/bandolier/loader/IResolver.java
@@ -13,11 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler.loader;
+package com.shapesecurity.bandolier.loader;
 
+import org.jetbrains.annotations.NotNull;
 
-public class ModuleLoaderException extends Exception {
-	public ModuleLoaderException(String module, Exception cause) {
-		super("Module Loader Exception: module " + module + " cannot be loaded: " + cause.getMessage(), cause);
-	}
+import java.nio.file.Path;
+
+public interface IResolver {
+	/**
+	 * Given a root and a (potentially relative) path, resolves the path to an absolute path.
+	 */
+	@NotNull
+	String resolve(@NotNull Path root, @NotNull String path);
 }
