+++ src/main/java/com/shapesecurity/bandolier/loader/ModuleLoaderException.java
@@ -13,19 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler.loader;
+package com.shapesecurity.bandolier.loader;
 
-import org.jetbrains.annotations.NotNull;
 
-import java.nio.file.Path;
-
-public class FileSystemResolver implements IResolver {
-	@NotNull
-	@Override
-	public String resolve(@NotNull Path root, @NotNull String path) {
-		if (path.startsWith(".")) {
-			return root.resolve(path).normalize().toString();
-		}
-		return path;
+public class ModuleLoaderException extends Exception {
+	public ModuleLoaderException(String module, Exception cause) {
+		super("Module Loader Exception: module " + module + " cannot be loaded: " + cause.getMessage(), cause);
 	}
 }
