+++ src/test/java/com/shapesecurity/bandolier/ImportExportPathRewriterTest.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler;
+package com.shapesecurity.bandolier;
 
-import com.shapesecurity.es6bundler.loader.FileSystemResolver;
+import com.shapesecurity.bandolier.loader.FileSystemResolver;
 import com.shapesecurity.shift.ast.Module;
 import com.shapesecurity.shift.codegen.CodeGen;
 import com.shapesecurity.shift.parser.JsError;
