+++ src/test/java/com/shapesecurity/bandolier/ImportExportTransformerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler;
+package com.shapesecurity.bandolier;
 
 import com.shapesecurity.shift.ast.Module;
 import com.shapesecurity.shift.codegen.CodeGen;
