+++ src/test/java/com/shapesecurity/bandolier/loader/NodeResolverTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.shapesecurity.es6bundler.loader;
+package com.shapesecurity.bandolier.loader;
 
 
 import org.jetbrains.annotations.NotNull;
